hydra/mpiexec: refactor the rest of the pmi handlers
Merge the rest of the PMI v1 and v2 handlers. Replace the function table
mechanism with direct if-switches.
